Split MembershipFactory into two classes: MembershipFactory and MembershipManager to follow Single responsibility principle. Added a TestMembershipManager class
